+++ src/config/mod.rs
@@ -1,3 +1,6 @@
+pub mod app;
+pub use app::AppConfig;
+
 use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};